allow updates within transaction to be dealt as update
update statements executed with autocommit False are being handled as insert statements. The results in scenario will not have a row count of rows updated. Interestingly updates that are being executed with autocommit True, go through the proper path and are treated as update statements and come back with valid update row counts.

Inline comments added for clarity to changes and adjacent blocks to help understand intent and usage. So the next person can read it quicker than myself.